# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  service.AICoreConfigurationRequest:
    properties:
      executableId:
        type: string
      inputArtifactBindings:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      name:
        type: string
      parameterBindings:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      scenarioId:
        type: string
    required:
    - executableId
    - name
    - scenarioId
    type: object
  service.AICoreConfigurationResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
  service.AICoreConfiguration:
    properties:
      createdAt:
        type: string
      executableId:
        type: string
      id:
        type: string
      name:
        type: string
      scenarioId:
        type: string
    type: object
  service.AICoreConfigurationsResponse:
    properties:
      count:
        type: integer
      resources:
        items:
          $ref: '#/definitions/service.AICoreConfiguration'
        type: array
    type: object
  service.AICoreDeployment:
    properties:
      configurationId:
        type: string
      createdAt:
        type: string
      deploymentUrl:
        type: string
      details:
        additionalProperties: true
        type: object
      id:
        type: string
      modifiedAt:
        type: string
      status:
        type: string
      statusMessage:
        type: string
    type: object
  service.AICoreDeploymentDeletionResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
  service.AICoreDeploymentDetailsResponse:
    properties:
      completionTime:
        type: string
      configurationId:
        type: string
      configurationName:
        type: string
      createdAt:
        type: string
      deploymentUrl:
        type: string
      details:
        additionalProperties: true
        type: object
      executableId:
        type: string
      id:
        type: string
      lastOperation:
        type: string
      latestRunningConfigurationId:
        type: string
      modifiedAt:
        type: string
      scenarioId:
        type: string
      startTime:
        type: string
      status:
        type: string
      statusDetails:
        additionalProperties: true
        type: object
      statusMessage:
        type: string
      submissionTime:
        type: string
      targetStatus:
        type: string
      ttl:
        type: string
    type: object
  service.AICoreDeploymentModificationRequest:
    properties:
      configurationId:
        type: string
      targetStatus:
        type: string
    type: object
  service.AICoreDeploymentModificationResponse:
    properties:
      deploymentUrl:
        type: string
      id:
        type: string
      message:
        type: string
      status:
        type: string
      targetStatus:
        type: string
    type: object
  service.AICoreDeploymentRequest:
    properties:
      configurationId:
        type: string
      ttl:
        type: string
    required:
    - configurationId
    type: object
  service.AICoreDeploymentResponse:
    properties:
      deploymentUrl:
        type: string
      id:
        type: string
      message:
        type: string
      status:
        type: string
      ttl:
        type: string
    type: object
  service.AICoreDeploymentsResponse:
    properties:
      count:
        type: integer
      deployments:
        items:
          $ref: '#/definitions/service.AICoreTeamDeployments'
        type: array
    type: object
  service.AICoreModel:
    properties:
      accessType:
        type: string
      allowedScenarios:
        items:
          additionalProperties: true
          type: object
        type: array
      description:
        type: string
      displayName:
        type: string
      executableId:
        type: string
      model:
        type: string
      provider:
        type: string
      versions:
        items:
          $ref: '#/definitions/service.AICoreModelVersion'
        type: array
    type: object
  service.AICoreModelVersion:
    properties:
      capabilities:
        items:
          type: string
        type: array
      contextLength:
        type: integer
      cost:
        items:
          additionalProperties: true
          type: object
        type: array
      deprecated:
        type: boolean
      inputTypes:
        items:
          type: string
        type: array
      isLatest:
        type: boolean
      metadata:
        items:
          additionalProperties: true
          type: object
        type: array
      name:
        type: string
      orchestrationCapabilities:
        items:
          type: string
        type: array
      retirementDate:
        type: string
      streamingSupported:
        type: boolean
      suggestedReplacements:
        items:
          type: string
        type: array
    type: object
  service.AICoreModelsResponse:
    properties:
      count:
        type: integer
      resources:
        items:
          $ref: '#/definitions/service.AICoreModel'
        type: array
    type: object
  service.AICoreTeamDeployments:
    properties:
      deployments:
        items:
          $ref: '#/definitions/service.AICoreDeployment'
        type: array
      team:
        type: string
    type: object
...
host: localhost:7008
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is the backend API for the Developer Portal, providing endpoints
    for managing organizations, teams, projects, components, and deployments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Developer Portal Backend API
  version: "1.0"
paths:
  /ai-core/configurations:
    get:
      consumes:
      - application/json
      description: Get all configurations from AI Core for the authenticated user's team
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved configurations
          schema:
            $ref: '#/definitions/service.AICoreConfigurationsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Core configurations
      tags:
      - ai-core
    post:
      consumes:
      - application/json
      description: Create a new configuration in AI Core for the authenticated user's
        team
      parameters:
      - description: Configuration data
        in: body
        name: configuration
        required: true
        schema:
          $ref: '#/definitions/service.AICoreConfigurationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created configuration
          schema:
            $ref: '#/definitions/service.AICoreConfigurationResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create AI Core configuration
      tags:
      - ai-core
  /ai-core/deployments:
    get:
      consumes:
      - application/json
      description: Get all deployments from AI Core for the authenticated user's team
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved deployments
          schema:
            $ref: '#/definitions/service.AICoreDeploymentsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Core deployments
      tags:
      - ai-core
    post:
      consumes:
      - application/json
      description: Create a new deployment in AI Core using a configuration
      parameters:
      - description: Deployment data
        in: body
        name: deployment
        required: true
        schema:
          $ref: '#/definitions/service.AICoreDeploymentRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Successfully scheduled deployment
          schema:
            $ref: '#/definitions/service.AICoreDeploymentResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create AI Core deployment
      tags:
      - ai-core
  /ai-core/deployments/{deploymentId}:
    delete:
      consumes:
      - application/json
      description: Mark deployment as deleted in AI Core
      parameters:
      - description: Deployment ID
        in: path
        name: deploymentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Successfully scheduled deployment deletion
          schema:
            $ref: '#/definitions/service.AICoreDeploymentDeletionResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete AI Core deployment
      tags:
      - ai-core
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific deployment from AI Core
      parameters:
      - description: Deployment ID
        in: path
        name: deploymentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved deployment details
          schema:
            $ref: '#/definitions/service.AICoreDeploymentDetailsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Core deployment details
      tags:
      - ai-core
    patch:
      consumes:
      - application/json
      description: Update target status or configuration of a deployment in AI Core
      parameters:
      - description: Deployment ID
        in: path
        name: deploymentId
        required: true
        type: string
      - description: Deployment modification data
        in: body
        name: modification
        required: true
        schema:
          $ref: '#/definitions/service.AICoreDeploymentModificationRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Successfully scheduled deployment modification
          schema:
            $ref: '#/definitions/service.AICoreDeploymentModificationResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update AI Core deployment
      tags:
      - ai-core
  /ai-core/models:
    get:
      consumes:
      - application/json
      description: Get all available models from AI Core for a specific scenario
      parameters:
      - description: Scenario ID to get models for
        in: query
        name: scenarioId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved models
          schema:
            $ref: '#/definitions/service.AICoreModelsResponse'
        "400":
          description: Bad request - missing scenarioId parameter
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: User not assigned to team or team credentials not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI Core models
      tags:
      - ai-core
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"